webadmin : Handle snapshot create multi-tab validation error
Snapshot create popup has 2 tabs :
1. General
2. Schedule
Some Editors in both tabs have validations for each input elements in
the tabs.However, although the validations validate as expected, the
problem was as under:
When user in some tab and a validation fails in an input element of
the other tab, although tab complained of error, it used to go
unnoticed as the tab reporting the error is not the one that is
active. This patch, activates the tab that has the errors in the
following order:
1. If the general tab has any validation failures, it opens up the
   general tab.
2. If the schedule tab has any failures, it opens up the schedule tab.
In accordance with above 2 steps, if both the tabs have validation
failures, immaterial of the number of failures in each tab or any
other criteria, general tab opens up(Does nothing if its already open).

Along with the above this patch also does the following:
1. Refactored EntityModelDateTimeBox and GwtDateTimeBox in accordance
   with most other widgets.
2. Corrected implementations of handling widget inValidity states.

Change-Id: I88c05aa369758ce174b0dc85f83a253e608d6c0e
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1215887
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1218606
Signed-off-by: Anmol Babu <anbabu@redhat.com>