restapi: Adjust display details when adding or updating VMs
A previous change modifed the way the display information is populated
to support the new concept of "graphics devices". But this change didn't
take into account that the display information needs to be adjusted for
backwards compatibility not only when retrieving VMs, but also when
adding or modifying them, as those operations also return the
representation of the VM. As a result when a VM is added or updated the
display type field isn't populated and that casuses issues for some
clients that expect this field. In particular this breaks the "rbovirt"
Ruby client. This patch changes the RESTAPI so that the display
adjustment is also performed when adding or updating VMs.

Change-Id: I35c8d28b9278294f780b073ee173520d501797f8
Related: https://bugzilla.redhat.com/1033547
Related: https://gerrit.ovirt.org/35281
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>