webadmin: locked vm icon
Until now, the in-memory lock status for a given VM was not exposed in
the UI. This patch changes the status icon to have an indication when
the VM is exclusively locked.

We add an icon of a lock which overlay the regular status icon of the
VM, in case it is exclusively locked.

Change-Id: I121c5e65eb3bdc740b1b0aaf05f5e12591a86e46
Signed-off-by: Arik Hadas <ahadas@redhat.com>