webadmin : Fix ClassCastException in datetimebox widget
This patch fixes the ClassCastException obtained from
recently introduced AssetProvider#addStateUpdateHandler().
The issue is with this patch it tries to cast the
GwtBootstrapDateTimePicker to widget and hence fails
this was fixed by making it extend from Composite
which in turn extends from Widget class and casting
from derived to base wouldn't be a problem.

Change-Id: Ib23eaab6160c814c26cd6a0566263cabed752e3e
Signed-off-by: Anmol Babu <anbabu@redhat.com>