core: Incorrect check of VirtIO-SCSI status on VM update
If VirtIO-SCSI device is to be enabled on VM update, cluster and OS
compatibility must be verified. If getParameters().isVirtioScsiEnabled()
is null, that means the current device status must be preserved. In this
case, cluster and OS compatibility must be verified, if VirtIO-SCSI
device is currently enabled.

This logic is correctly implemented in
UpdateVmCommand.isVirtioScsiEnabled() method, so we use it instead of
incorrect condition used previously.

Change-Id: I459256da7df4ef0f6662479df2d6d0a6cabc001b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1219697
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>