webadmin: Enable editing name on import iscsi SD
The import of a storage domain will fail if an existing domain has the
same name.

For block storage domains, the name of the storage domain in the table
in the import domain window of will be now editable.
The user can change the name before importing the domain.

Note that the new name will not be updated in the storage domain
metadata unless the user will change the name when the storage
domain will be active.
It is the same behavior as in REST API.

Change-Id: I70dcd7432ad3e5fb9f8bd978e87dbe777dabde18
Bug-Url: https://bugzilla.redhat.com/1195784
Signed-off-by: Fred Rolland <frolland@redhat.com>