webadmin: Tooltip in Disks subtab shows Original Allocation Policy
1. Select a VM in Virtual Machines list.
2. Select Disks subtab.
3. Select Images radio button.

Allocation Policy column in the disks table shows current allocation
policy of the disk. If snapshots were created, the current allocation
policy will be always "Thin provision", while allocation policy of the
original disk may be different.

To make the original allocation policy easily accessible, tooltip was
added to the Allocation Policy column. The original allocation policy is
displayed in the tooltip.

Change-Id: I267eb950c57f4c5cca92d5fdd04e1294b1b2f2b0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1117003
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>