core: VM compatibility version scheduler filter
This patch is part of the series introducing VM compatibility version
feature. This feature allows to customize a VM to its own compatibility
version unrelated to one of the cluster. VM compatibility version is
limited by Data Center and capabilities of the host the VM will be able
to run on.

This patch adds a new filter to the scheduler to filter out hosts that
do not support the desired VM compatibility version.

Change-Id: I96967b10e3af8f31f81c16e461ef69d5fb605833
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>