core: make postConstruct final
Consider the following case:
Class A {
  @PostConstruct
  protected void foo() {}
}
Class B extends A {
  @Override
  @PostConstruct
  protected void foo() {}
}

JBOSS 7.1.1: when instantiate class B, B#foo will be called twice. This
seems to be a bug in JBOSS.
JBOSS EAP 6.3: when instantiate class B, B#foo will be called only once.

Because of Ibf9cdd60, we have @PostConstruct on postConstruct methods in
commands as well as in the CommandBase, so we have the structure above,
thus in JBOSS 7.1.1 the postConstruct methods are called multiple times.

This patch introduces a different approach where the postConstruct
method that is annotated with @PostConstruct is declared as final in
CommandBase, and calls CommandBase#init which is the method that should
be overridden by commands and does not need to be annotated with
@PostConstruct.

Change-Id: I8340d8348b298fec0c73bf93b7411a149e9fe8c0
Signed-off-by: Arik Hadas <ahadas@redhat.com>