core: fix compensation for import and clone vm
If the initialization that is done in post-construct accessed the
command's parameters in a non-transactive command that uses
compensation, the compensation didn't work. The problem is that when
instantiate the command for compensation, the parameters are not passed,
so the initialization in the post-construct failed.

The solution is to do the initialization in post-construct only when the
parameters are set.

Change-Id: Ie0e6d4b67d09537dedd167937e8949e7875d40a2
Signed-off-by: Arik Hadas <ahadas@redhat.com>