userportal, webadmin: include jQuery 1.11.3
We explicitly depend on jQuery now, but were previously relying
on PatternFly's copy of it. Since it is an explicit dependency,
we should include it ourselves.

Also -- older versions of PatternFly (<1.2.1) shipped a 2.0+
version of jQuery. When oVirt included this version, IE8 could
not render, since jQuery 2.0 drops IE8 support. While we will
no longer support IE8 in oVirt 3.6.0+, it's nice to have it at
least render.

Change-Id: I32b24168795e1852a3ebb142a7b6c364c9808107
Bug-Url: https://bugzilla.redhat.com/1222584
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>