core: Fix FB issue in VdsCpuUsageComparator
This patch fixes the following FindBugs warning on this class:

"""
Se: Comparator doesn't implement Serializable
(SE_COMPARATOR_SHOULD_BE_SERIALIZABLE)

This class implements the Comparator interface. You should consider
whether or not it should also implement the Serializable interface. If a
comparator is used to construct an ordered collection such as a TreeMap,
then the TreeMap will be serializable only if the comparator is also
serializable. As most comparators have little or no state, making them
serializable is generally easy and good defensive programming.
"""

Change-Id: Iedbed23022e170c5b9486b4af26ac9ed1eec96eb
Signed-off-by: Allon Mureinik <amureini@redhat.com>