core: fix creation of instance-type
The addition of cinder disks broke creation of instance-type. While
trying to create an instance-type, one would have get NPE in the
AddVmTemplateCommand where cinder disks were about to be added.

The problem was that instance-types are not based on any existing VM so
getVm() returns null. The solution is to check if getVm()!=null before
trying to add cinder disks. This is ok because instance-type don't have
disks and in particular don't have cinder disks.

Change-Id: Idb8b7e439f675fcfe29f403770a5ca52357ab6c4
Signed-off-by: Arik Hadas <ahadas@redhat.com>