webadmin : Fix volume main and sub tab column id issue
This patch does the following,
1. Adding column's to MainTabVolumeView require columns to override
AbstractColumn or handle id on own. But overriding AbstractColumn
provides base implementation for ColumnWithElementId#configureElementId
which actually assigns ids for columns.
Accordingly changes are performed in,
   [i] Changed VolumeActivityColumn
2. Now, ApplicationTemplates#statusTemplate takes a SafeHtml and a
string id. Some of the cells corrected as part of this patch used to
treat tooltips as ids and now this patch corrects it by adding new
templates to ApplicationTemplate for handling tooltip requiring images
and also handles ids for such columns.
Accordingly changes are performed in,
   [i] GlusterCapacityCell and hence VolumeCapacityCell and
       BrickCapacityCell
  [ii] VolumeInfoCell

Change-Id: Ifded690b427ba8eb17bba272360f916850f76891
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1220393
Signed-off-by: Anmol Babu <anbabu@redhat.com>