services, setup: ovirt-vmconsole integration
integration between ovirt-engine and ovirt-vmconsole-proxy.

Includes:
- helpers to integrate ovirt-engine and ovirt-vmconsole-proxy
- setup plugins for PKI artifacts

Key assumption: ovirt-vmconsole-proxy-helper will run on the
same host of ovirt-engine.

The setup code intentionally does the bulk of work in the
engine directories, and copy the configuration/pki files
only as last step. Doing so, if for any reason the proper
directory is not accessible by setup, the correct configuration
is created anyway, and the admin has just to copy it
with the right permissions.

When the helper is actually run through ovirt-vmconsole-proxy,
selinux prevents its execution. This will be addressed
by a followup patch.

Feature wiki page: http://www.ovirt.org/Features/Serial_Console

Bug-Url: https://bugzilla.redhat.com/1223671
Change-Id: I034ef8e6d10da5dc93eda61e0c5c518ca13a5a28
Signed-off-by: Vitor de Lima <vdelima@redhat.com>
Signed-off-by: Francesco Romani <fromani@redhat.com>