core: configurable parent for create image template
We used to assume that CreateImageTemplateCommand is called from
AddVmTemplateCommand, so the parent command was hard-coded in
CreateImageTemplateCommand.

This patch modifies CreateImageTemplateCommand such that it will take
the parent command from the parameters instead. This change is needed
for adding commands that extend AddVmTemplateCommand such as add
template from snapshot.

Change-Id: If81b0e84e32674c094a2fbeaf1a2e26f2296a7f3
Bug-Url: https://bugzilla.redhat.com/1003492
Signed-off-by: Arik Hadas <ahadas@redhat.com>