gluster: Corrected geo-rep case of volume snapshot creation
While creating gluster volume snapshot for a geo-replicated
volume, we try to pause the geo-rep session and then create
snapshot for the slave volume first. The pause of the slave
volume should be done only if geo-rep session is not in
one of the states from CREATED, PAUSED or STOPPED.
Modified the code accordingly.

Change-Id: Iaefc24c83168ce53960c9f90597a43d5030f7212
Bug-Url: https://bugzilla.redhat.com/1224902
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>