core: add timestamp for vm-dynamic updates
added a timestamp in VmManager that save the time vm information
updated.
this is updated by consumers of the VmManager.lock object,
and used by VmsMonitoring, which gets vm data at a time and lock at a
different time - using this timestamp, monitoring cycle of 'dirty' vm
can be skipped.

Change-Id: If858993450cf957babdcf337e9b7867a282657bc
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>