webadmin : Fix Menu Cell click event
A cell registers to events that it wants to listen
with its table. The table notifies the cell by invoking
onBrowserEvent when a registered event occurs.

In case of ActivityColumn of Volumes main tab, it is a
column of a CompositeCell - VolumeActivityCompositeCell
which is in-turn a combination of :

1.VolumeActivityStatusColumn - Column of AbstractCell
                               VolumeActivityStatusCell
2.Column of AbstractCell - VolumeActivitySeperatorCell
3.Column of AbstractCell - MenuCell - for rebalance
4.Column of AbstractCell - MenuCell - for remove-brick

So, this being the situation, the Composite of the Composites
which is the VolumeActivityCompositeCell receieves the events
to which it subscribes from the table and then only constituent
cells can get them but in our case, the VolumeActivityCompositeCell
was not subscribing for the click event and hence the constituent
MenuCell was not notified of the click.
Hence this patch modifies VolumeActivityCompositeCell to
subscribe to a superset of all events that its constituent cells
need.

Change-Id: I32a223494d23848c530ee083dc14e1d072a86967
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1209422
Signed-off-by: Anmol Babu <anbabu@redhat.com>