restapi: Move Cluster remove from collection to entity
Currently the methods implementing the DELETE operation are part of the
collection interfaces. This causes an issue with recent version of
Resteasy that implement the JAX-RS specification strictly. The issue is
that the remove resource method and the subresource locator use the same
URI pattern "/cluster/{id}". According to the JAX-RS specification when
two methods use the same URI pattern the resource method should be used
and the resource locator should be ignored. See section 3.7.2, step 2.h
of the resource matching algorithm. To avoid this problem this patch
moves the remove method from the collection resource to the entity
resource, but only for the cluster entity. Subsequent patches will move
all the other remove methods.

Change-Id: Ife8a3eb08ab379ffc0ea7b41c49f164cb55c15f2
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>