db:set AutoRegistrationDefaultVdsGroupID correctly
This patch fixes an issue in post-upgrade when attempting to set
AutoRegistrationDefaultVdsGroupID in vdc_options to the value of the
default cluster fails and returns null when the default cluster name was
changed from 'Default' to something else.

In that case we will set the configuration value to the first cluster id
we can find.

Change-Id: If1aa05a0d24866dfad3a9d18d90262ec07d2fdac
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1226944
Signed-off-by: emesika <emesika@redhat.com>