engine: Block provide type change
There is no sense to modify the provider type, i.e. when
an OpenStack provider is created, the ovirt-engine creates
the required storage domain for it. Modifying the type of
a provider to a different type modifies the entire nature
of the provider.

Each provider has its own definition based on its type and
modifying the type will not be supported.

Change-Id: I3f5b0980977bc9fe556b1e3a72b1942bd8d86cdc
Bug-Url: https://bugzilla.redhat.com/1227320
Signed-off-by: Moti Asayag <masayag@redhat.com>