restapi: Move Device remove from collection to entity
This patch moves the method that implements the DELETE operation from
the collection interface to the entity interface. This is needed to
avoid issues with newer versions of Resteasy.

For most entities the method can be moved with simple changes, but in
this case there are many related classes and tests, and moving the
method requires a lot of changes. To avoid possible regressions and to
not block the migration to WildFly this patch solves the issue moving
the method only in the interfaces, and introducing a new method in the
entity resources that calls the existing method of the collection
resources. This way there is no need to do complicated changes to the
resources and the tests. This is only a temporary solution, in the
future the method will be actually moved.

Change-Id: Ie41693cde8cdb2cb82cbc039f65dd540607629ca
Related: https://gerrit.ovirt.org/41783
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>