core: Introduce volume classification
Introduce new column in images table of volume_classification.
volume_classification is a column which determines if the image is a
volume or a snapshot.
It should be mainly used for Cinder, since a previewed snapshot in
Cinder is acctually a new Volume in Cinder and this indication is
needed once the disk is being deleted.

Change-Id: Ia1f37abcc4c5a99076138d7dfb0ffe72f68aef50
Bug-Url: https://bugzilla.redhat.com/1185826
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>