core: do not restrict user public keys to RSA
The users can upload ssh public keys, to be used to perform
authentication in to the ovirt-vmconsole-proxy, and then to
connect to virtio serial console.

For early implementation reasons, the key type was restricted to
RSA. This patch lift the restrictions, allowing the user to
upload any valid SSH key.

The code should now perform only the basic sanity check.

Change-Id: If3b677a140cdedd478856bf8aa62ac758932767e
Signed-off-by: Francesco Romani <fromani@redhat.com>