backend: Enable granular hostdev passthrough support
Added flag to VmHostDevice entity that marks whether
this device should be visible to the guest OS (iommuPlaceholder=false)
or is there just to satisfy IOMMU group constrained for other (primary)
device (iommuPlaceholder=true).

This flag is send to VDSM as a part of spec params.

When adding/removing VmHostDevices we take this flag into account.

Change-Id: I1c8a80adce97ef98174bd85ac75f09dfe2a36de9
Signed-off-by: Martin Betak <mbetak@redhat.com>