webadmin: it is possible to attach the same disk twice
Problem happend in the "instance image flow" only. The problem was that the
disks which can not be attached (e.g. the ones already attached) were returned
only from server. But in the instance image flow the disks are first configured
in the new/edit vm dialog and than submitted at once.

Fixed by removing the this disks from the list of disks to be attached.

Change-Id: I3eb3b0608e957c88c98c264e87dbb651b179c8e3
Bug-Url: https://bugzilla.redhat.com/1228975
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>