core: Auto-Import the hosted engine VM if its not managed
Let the monitoring import the HE VM if its not in the setup already

PRECONDITIONS
- The hosted engine Storage Domain (name Hosted_Engine) should be
  imported into the setup and be Active
- The engine VM should already be in the DB. The Vm is an external Vm
  at first and assumed to be added by the monitoring of the host

IMPORT
The import rely on the "full list" response from vdsm to get all
data that is out there. the disks are assumed to be on the target storage domain.

NEW
OriginType.MANAGED_HOSTED_ENGINE was added to mark that VM as
managed from now on, and to signal the monitoring on when its not imported
and should be.

Change-Id: I0ff2c440785cb91e3beef0530eda77324cc954d0
Bug-Url: https://bugzilla.redhat.com/1160094
Signed-off-by: Roy Golan <rgolan@redhat.com>