restapi: Move VM Disk remove from collection to entity
Previous changes moved the remove methods from collection to entity
interfaces, except the device related ones as those required larger
changes that posed a risk for the migration to WildFly, those were
handled in a different and less nice way. This completes the movement
for the VM disk resource in a nicer way.

Change-Id: I72f35d4a2ed77b249ec142c7b27782694e312efc
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>