core: SPICE copyPaste/fileTransfer not passed properly
The problem was that the VmInfoBuilderBase checked if the SPICE device is there
according to the graphicsInfo.
The problem was that the graphicsInfo is a dynamic attribute containing things
like IP, port etc which exists only after the VM has been created - here we are
just creating it so it does not yet exist.

Fixed by creating this information in the RunVm and RunVmOnce commands
according to the DB value/parameters passed

Change-Id: Ic09709b579fd3e2fe622790f66286537ad01e728
Bug-Url: https://bugzilla.redhat.com/983019
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>