core: Make regular user owner of VMs he created
If regular user has permission to create a VM and does this through
webadmin UI (having also ReadOnlyAdmin role) or REST API with
Filter: false, he doesn't get ownership for this VM by default. This is
OK for admin users, but non-admin users will effectively loose access to
their VMs.

To fix this, additional check is made for webadmin UI users and REST API
with Filter: false users, checking if they have permission to manage the
VM being created. If not, this permission is added for them.

Change-Id: I7e814ab024ceed63e3ec7fce82ca574fa5be1c8b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1232419
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>