uicommonweb: Attach virtual IDE disks new behavior
When a user tries to hot plug an IDE disk, he gets an error message
after the operation completes, telling him that an IDE disk cannot be
hot plugged. However, the disk is also not attached to the VM.

This patch introduces a new approach:
If the "Activate" check box is checked, a warning is shown if at least
one IDE disk is selected from the Images/LUNs list. The warning tells
the user that the disk will be attached but not activated, and that he
should activate manually after the VM is shut down.
In addition, a new tooltip was added to the "Activate" check box,
explaining that only non IDE disks are activated after attached.

Change-Id: I187c74c7c39f438dfa5e81942f4a6aec41323830
Bug-Url: https://bugzilla.redhat.com/1167248
Signed-off-by: Idan Shaby <ishaby@redhat.com>
Signed-off-by: Tal Nisan <tnisan@redhat.com>