core: move vdsm timestamp to VmInternalData
Move the timestamp we get from VDSM on List/VmStats/AllVmStats to be in
VmInternalData instead of VmDynamic. This information does not belong to
VmDynamic since it neither describes a dynamic property of a VM nor
something else about the properties of the VM. This field represents
kind of meta-data about the message that was received.

Change-Id: I1a1f37ca4ce4d0af0ff05bc7531074e3adfafa4e
Signed-off-by: Arik Hadas <ahadas@redhat.com>