webadmin: Check for VM status when activating IDE disks
When the VM is down and one tries to attach and activate an IDE disk, he
gets an error message that tells him he can't activate it and that he
should turn off the VM first. In addition, the disk is attached but not
activated although the VM is down.
This patch adds a check for the VM status to fix this issue.

Change-Id: Ic2fe5e5974890ee591f6a903d9ea9ae024a22ac3
Bug-Url: https://bugzilla.redhat.com/1167248
Signed-off-by: Idan Shaby <ishaby@redhat.com>