core: dont check vmManager timestamp if it wasnt set
when engine starts, the first monitoring cycle will always skip the
current running vms because the VmManager objects for them are created
only after fetching the information from vdsm, so the time check will
always fail.

the check is changed to verify that some value has been set,
and if not it considered ok to do the monitoring cycle.

Change-Id: I6ced6416b4015b86aeb75371224ad68dc1f5b74c
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>