core: send maxMemory and maxMemSlots only if hotplug memory is supported
in case new vdsm runs in old (< 3.6) cluster,
the engine will send the new maxMemory and maxMemorySlots parameters on
run vm, but these are not supported with old emulated machines,
and will cause the run to fail.

sending the new parameters only if hotplug memory is supported in the
cluster.

Change-Id: I194a4eee3736d0de842fcf3474a4f6f0580b6577
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>