webadmin: Forbid hot plug availability when VM is suspended
When VM is on suspended status, the "Activate" and "Deactivate" buttons
under Virtual Machines -> Disks were available for VirtIO disks and
could be clicked.
Since we don't support disk hot plug while VM's are in suspended status,
and since it fails in the CDA anyway, this patch changes these buttons
availability in this case.

Change-Id: Ia2d610c3dda4445f6a071ba23a2b6abf6699fbc9
Bug-Url: https://bugzilla.redhat.com/1234098
Signed-off-by: Idan Shaby <ishaby@redhat.com>