core: TestingAffinityRulesEnforcementPerCluster
Previously, this test used an anonymous inner extension of
AffinityRulesEnforcementPerCluster that assigned mocks to its  DAOs in
the wakeup() method. This behavior caused a NullPointerException since
the DAOs were also used elsewhere.

This method extracts the anonymous class to a named inner class to allow
setting the DAOs in the constructor and use throughout the class' life.

Change-Id: I2b722fb78d1ec32652a0315e1fcc88ebd46ff8e1
Signed-off-by: Allon Mureinik <amureini@redhat.com>