packaging: sercon: fix copying of artifacts
Outside developer mode, it is wrong to copy artifacts
in the same transaction who creates them.
The copy will (obviously) fail, and no artifact at all
will be created.

It is better to create the artifact we need directly
in the final destination, and copy the others. We are
already in a trasaction, so this is safe as the previous
design, and proabably less confusing.

Change-Id: I5a8d7a4edb92a872f37f89a7aef1187b1a03a542
Signed-off-by: Francesco Romani <fromani@redhat.com>