restapi: Fix metadata of "refreshluns"
The patch that introduced the "refreshluns" actions also explicitly
documented the "async" matrix parameter for this new action and for the
existing "isattached" action. This isn't correct, as the "async"
parameter is implicitly documented, and documenting it explicitly causes
a duplication. That duplication induces an error in the code generated
by the Java SDK. This patch removes that explicit documentation and also
removes empty elements from the RSDL metadata.

Change-Id: I402ff58e93758e7593c190245cf0a6184af285da
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>