core: separate iteration count for vms and host monitoring
Host and VMs monitoring are now two different quartz jobs. It does not
makes sense that they will use the same counter to count their
iterations.

This patch adds new counter in VdsManager for counting the host
monitoring iterations. This counter works exactly as the counter in
VMStatsRefresher worked before. In addition the counting of iterations
of VMs monitoring is simplified.

Change-Id: I80e37c501de4f4117400e21c6ceef86911515c89
Signed-off-by: Arik Hadas <ahadas@redhat.com>