core: make PollVMStatsRefresher abstract
Better class hierarchy for poll-based vm-stats refreshers is to have an
abstract class that contains the common things for polling and then
specific refreshers will extend it.

This patch changes PollVMStatsRefresher to be abstract and presents a
new refresher PollListAndAllVmStatsRefresher that extends it and
implements the legacy polling that we used to have before 3.6.

In addition, PollAllVmStatsOnlyRefresher is changed to implement only
the things that varies instead of override the 'poll' method.

Change-Id: I2d6277d3ca0ab14a9431284a13744943b6a4945b
Signed-off-by: Arik Hadas <ahadas@redhat.com>