restapi: Remove "performRemove" method
Now that the remove methods have been moved from the collection to the
entity interfaces this method is no longer needed, except for an edge
case in the device resource tests.  This patch removes it except from
almost all places, except those where the tests need it. Later patches
will change the tests so that the method can be completely removed.

Change-Id: Ic2eaec165059e9501940cb427edcaeabb6366ed2
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>