restapi: Move CDROM remove from collection to entity
Previous changes moved the remove methods from collection to entity
interfaces, except the device related ones as those required larger
changes that posed a risk for the migration to WildFly, those were
handled in a different and less nice way. This completes the movement
for the CDROM resource in a nicer way.

Change-Id: Iab2ad3729844d2697e6162df23e7ad34b19342bb
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>