frontend: Make EntityModelCellTable Sortable
Added missing support for sorting of host devices in the
AddVmHostDevicePopupView.

It was necessary to extract the sort handler logic from AbstractActionTable to
ColumnResizeCellTable in order for it to be accessible in EntityModelCellTable.

The usage of SortedListModel interface also required adding a new generic
parameter to SplitTable and it's descendants 'L' specifying exact subclass of
ListModel being used.

Change-Id: Ia64f57af66b3eb8bd0f69e830b2f53e7f56badc2
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>