restapi: Move Watchdog remove from collection to entity
Previous changes moved the remove methods from collection to entity
interfaces, except the device related ones as those required larger
changes that posed a risk for the migration to WildFly, those were
handled in a different and less nice way. This completes the
movement for the Watchdog resource in a nicer way.

Change-Id: I93c8c32eff508bd33f16b933069dfe54798d63dc
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>