restapi: Move VM NIC remove from collection to entity
Previous changes moved the remove methods from collection to entity
interfaces, except the device related ones as those required larger
changes that posed a risk for the migration to WildFly, those were
handled in a different and less nice way. This completes the
movement for the VM NIC resource in a nicer way.

Change-Id: I2df187c2933944a4694d62914d2babf291e0a628
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>