restapi: Move Template NIC remove from collection to entity
Previous changes moved the remove methods from collection to entity
interfaces, except the device related ones as those required larger
changes that posed a risk for the migration to WildFly, those were
handled in a different and less nice way. This completes the
movement for the Template NIC resource in a nicer way.

Change-Id: I30bb85bdf978e35d5d43b992181a42e33cf1daef
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>