engine: VM.getCompatibilityVersion() usage
This patch is part of the series introducing VM compatibility version
feature. This feature allows to customize a VM to its own compatibility
version unrelated to one of the cluster. VM compatibility version is
limited by Data Center and capabilities of the host the VM will be able
to run on.

This patch makes using VM.getCompatibilityVersion() instead of
VM.getClusterCompatibilityVersion() where appropriate.

Change-Id: I4d8b71f38ca4a33c340922b06d4bf9b3e2b3c82b
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>