restapi: Remove @SingleEntityResource
This annotation was used to mark resource locators that return entity
resources, to distinguish them from actions and sub-collections. The
annotation was used by the "getEntity" that uses reflection to call the
lcoator and then the "get" method to retrieve the entity representation.
This "getEntity" method was in turn used by the "performRemove" method
in order to check that the entity exists before trying to remove it. All
these things aren't needed now because the remove methods are now part
of the entity resources, and they can directly call the "get" method
without these mechanism.

Change-Id: Ie98fbc2485aa5ffa946c4f45e08d91f9112cb893
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>