webadmin: improve the layout of import vm dialog
Change the background and the border of the top panel that contains the
selection of the source input of the import operation to highlight that
it is some kind of header for the dialog.

In addition, we now use fixed height for the panel at the center of
the dialog.

Change-Id: Ib3cb05e56c85f52e775f9cc60c3bf33dd0eda772
Signed-off-by: Arik Hadas <ahadas@redhat.com>