build: remove commons-io from module overlay
When trying to test a Cinder openstack provider,
we get an error of no such method:
  java.lang.NoSuchMethodError:
  org.apache.commons.io.output.DeferredFileOutputStream.
  <init>(ILjava/lang/String;Ljava/lang/String;Ljava/io/File

It appears that since we upgraded to wildfly, resteasy tries to call
the new constructor of DeferredFileOutputStream which was introduced
only at commons-io 1.4.
The following patch changes the pom.xml to use the default commons-io version,
the same as it is supported with jboss, and also removed all relation to
commons-io from the backend manager dependencies.

Change-Id: I18b3cb941338ef6519ed15ea7a428ff6636987d2
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>