core: reduce coupling between VmAnalyzer and VdsManager
VmAnalyzer was used to ask VdsManager if we are in a refresh iteration
to determine if VM statistics should be updated or not. This patch
changes that with a flag that is now being passed to VmsMonitoring that
specifies whether VM statistics should be updated or not, and VmAnalyzer
now check this flag instead.

Change-Id: I12a3608c5d8d6bce8c3ffe8b015ac81b57509780
Signed-off-by: Arik Hadas <ahadas@redhat.com>