restapi: The "isattached" action should return an action object
Currently the "isattached" action of the storage domain resource returns
a storage domain object. For example, the following HTTP request:

  POST /api/storagedomains/{storagedomain:id}/isattached
  <action>
    <host id="{host:id}"/>
  </action>

Returns the following response:

  <storage_domain id="{storagedomain:id}">
    <is_attached>true</is_attached>
  </storage_domain>

This isn't correct, because all actions are expected to return an action
object. Returning something different causes issues in the SDKs. The
returned result should look like this:

  <action>
    <status>
      <state>complete</state>
    </status>
    <is_attached>true</is_attached>
  </action>

There is no need to return the storage domain object, only the
"is_attached" flag, and it should be an attribute of the action object,
not of the storage domain.

This patch modifies the storage domain resource so that this action
works as described above.

Change-Id: I4033daddfca8f6d25ee2a1efd03fa2b65a48c6f4
Bug-Url: https://bugzilla.redhat.com/1234937
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>