core: display type was not possible to set from REST
For template/instance type it was not possible to change the graphics devices
(e.g. display->type in REST) since REST sends the defaultDisplayType as null so
the engine will pick it.
This logic of picking was implemented for new/edit VM but not for new/edit
template or instance type.

Change-Id: Ifbc5f2e4984a0f668c6de7ff1ebaa47699131088
Bug-Url: https://bugzilla.redhat.com/1236497
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>