restapi: Explicitly use Xalan as transformer
Currently the FOP message body writer used to generate PDF document
doesn't explicitly request a particular implementation of the XSLT
processor. When running in the WildFly application server this results
in an exception because an unresolved issue:

  Xalan Linkage error : TransformerConfigurationException
  https://issues.jboss.org/browse/WFCORE-519

To avoid that issue this patch changes the FOP message body writer so
that it explicitly requests the Xalan processor included in the
"org.apache.xalan" module.

Change-Id: I1c8cca9cc83e29a98d026c3226ab0f11c40660d3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>