webadmin: move activate storage domain checkbox down
In the import SD dialog, "Activate Domain in Data Center" checkbox was
moved down to the end of the view, since
A. It never really belonged up there (in the upper section).
B. Once the missing fields (name, description, comment) are restored
in a future patch, there's no room for this checkbox in the upper
section.

Change-Id: Iec242b302e3659c7b3f0da4f12eda3dfce5b1fc9
Signed-off-by: Vered Volansky <vvolansk@redhat.com>