aaa: Honor 'Session-TTL' header
Currently API requests may have "Session-TTL=xxx" header on them.
This should set the 'soft-limit' (in minutes) for this session in
the engine, but in fact is ignored.

This patch causes this header to be honored. For requests that open
a new session (meaning: cause a log-in), the value of the header is
saved and used later every time the session is refreshed (instead of
the default ConfigValues.UserSessionTimeOutInterval)

Bug-Url: http://bugzilla.redhat.com/1157751
Change-Id: If6d01d1afc2045ee1ed267512e8d5755a6413d4f
Signed-off-by: Ori Liel <oliel@redhat.com>