webadmin: Remove the "selectedItem" terminology from StorageModel.
privateAvailableiStorageItems List now appears only behind the scenes. We have
StorageType and StorageDomainType lists seperatly now, which *are*
displayed. Though this list is no longer displayed, it plays a great logical
part. This patch removes the list.
This is the first setp in making StorageModel inherit from Model instead
of ListModel<IStorageModel> .

Change-Id: I5d16972f8155ccea4505fd268683cbe92b079a84
Signed-off-by: Vered Volansky <vvolansk@redhat.com>