host-deploy: apply user visible messages logic
remove the use of InstallerMessages, this class is now being used only
for node upgrade and should be removed when that class is migrated to
the ovirt-node-upgrade interface.

VdsMgmtPacakges issues user visible messages only if it is not in check
only mode.

VdsDeploy is interacting with user for all messages as in current
implementation, the double log/audit was removed in favour of audit
only, as it already writes to log.

Change-Id: I6480d5d2d43caccc6093300e03a840afca5e899d
Signed-off-by: Alon Bar-Lev <alonbl@redhat.com>