restapi: Add 'placeholder' attribute to VmHostDevice resources
So far the backend only returned the devices that are to
be attached to the VM. Now it returns all host devices
configured for the VM and distinguishes IOMMU placeholder
devices with additional boolean flag.

This patch exposes to the REST users that some host device
is configured on the VM only as a placeholder using the
optional <placeholder>true</placeholder> element.

Change-Id: I76e9945cb680537a83b036b6f87cf5aeb33fa420
Signed-off-by: Martin Betak <mbetak@redhat.com>