core: dont update devices if vm is not updated
vms monitoring should skip updating devices of vm that is not marked for
update. if devices-hash changes, the vm should be in the vmDynamicToSave
list, because the analyzer will identify the hash is different.
this also ensures that the VmManager lock is taken for this vm and its
data was not invalidated by other command.

Change-Id: Idc60abd6ba072fe202f1cd50de2879800a4a0ee8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1230082
Signed-off-by: Omer Frenkel <ofrenkel@redhat.com>