core: handle Live Merge when top volume is larger than base
When a Live Merge is perfomed today, it internally performs a block
commit operation to copy data from a volume to its parent.  If the disk
had been extended after the snapshot was created, the sizes of the two
volumes will not match.

In most cases the size is adjusted properly by the underlying processes
performing the merge; to account for this, engine will update the size
of the volume in the database to reflect the new, correct size.

In the case of a raw volume on block storage, an explicit extension
needs to take place before the merge.  For this, a new child command of
Live Merge has been added to run before the merge starts.

For further detail, see comments to:
  https://bugzilla.redhat.com/show_bug.cgi?id=1232481

Change-Id: I628323b799a37691c40ea97590d4967f3ebf4a2b
Bug-Url: https://bugzilla.redhat.com/1241433
Signed-off-by: Greg Padgett <gpadgett@redhat.com>