webadmin: no way to set instance type
When selected an instance type it was reset to "custom" immediately.

The regression has been introduced by https://gerrit.ovirt.org/#/c/41486/2 by
call getListModel().setSelectedItem(null); from EntityModelCellTable which
fires the events outside of the control of the models.

Since this broke the VM dialog already second time the fix was to rewrite the
priority to use list box instead of EntityModelCellTable.

This patch also adds the callback to PriorityUtil so the caller can
activate/deactivate the instance type manager as needed.

Last change was to change the watchdog panel so it looks consistently with the
rest of the dialog (the difference became obvious after changing the priority
to list box).

Change-Id: I6127fb69c8b201e56ab1f13e28448c9fdebc4196
Bug-Url: https://bugzilla.redhat.com/1241403
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>